+++ CHANGES.rst
@@ -1,6 +1,15 @@
 Changelog for Rapid Photo Downloader
 ====================================
 
+0.9.0a7 (2017-01-xx)
+--------------------
+
+ - Disallow the running of the program as the root user.
+
+ - Updated program installer to stop installation on Fedora 23 and Debian Jessie.
+
+ - Corrected error in Spanish translation that caused crash when download started.
+
 0.9.0a6 (2016-12-10)
 --------------------